handle cases with no mutations found

+++ tasks/gene_typing/drug_resistance/task_chroquetas.wdl
@@ -50,29 +50,41 @@ task chroquetas {
   
     # rename output files to include sample name
     based_name=$(echo $(basename ~{genome_fasta}) | sed -E 's/(.*)\.[^\.]+$/\1/')
-    mv chroquetas_out/${based_name}.ChroQueTaS.AMR_stats.txt chroquetas_out/~{samplename}.ChroQueTaS.AMR_stats.txt
-    mv chroquetas_out/${based_name}.ChroQueTaS.AMR_summary.txt chroquetas_out/~{samplename}.ChroQueTaS.AMR_summary.txt
+    if [ -f chroquetas_out/${based_name}.ChroQueTaS.AMR_stats.txt ]; then
+      mv chroquetas_out/${based_name}.ChroQueTaS.AMR_stats.txt chroquetas_out/~{samplename}.ChroQueTaS.AMR_stats.txt
+      mv chroquetas_out/${based_name}.ChroQueTaS.AMR_summary.txt chroquetas_out/~{samplename}.ChroQueTaS.AMR_summary.txt
+    
+      # extract AMR summary string
+      # e.g. <GENE>_<REF_POSITION><AA_CHANGE><QUERY_POSITION>
+      tail -n+2 chroquetas_out/~{samplename}.ChroQueTaS.AMR_summary.txt \
+        | awk '{ print $1, $4, $3, $5 }' \
+        | sed -E 's/([^ ]+) ([^ ]+) ([^ ]+) ([^ ]+)/\1_\2\3\4/' \
+        | tr '\n' ',' \
+        | sed -E 's/,$//' \
+        | tee AMR_SUMMARY_STRING
   
-    # extract AMR summary string
-    # e.g. <GENE>_<REF_POSITION><AA_CHANGE><QUERY_POSITION>
-    tail -n+2 chroquetas_out/~{samplename}.ChroQueTaS.AMR_summary.txt \
-      | awk '{ print $1, $4, $3, $5 }' \
-      | sed -E 's/([^ ]+) ([^ ]+) ([^ ]+) ([^ ]+)/\1_\2\3\4/' \
-      | tr '\n' ',' \
-      | sed -E 's/,$//' \
-      | tee AMR_SUMMARY_STRING
-
-    # extract AMR summary annotated w/fungicide resistance
-    # e.g. <GENE>_<REF_POSITION><AA_CHANGE><QUERY_POSITION>[<FUNGICIDE_RESISTANCE1>;<FUNGICIDE_RESISTANCEn>]
-    tail -n+2 chroquetas_out/~{samplename}.ChroQueTaS.AMR_summary.txt \
-      | awk '{ print $1, $4, $3, $5, $6 }' \
-      | sed -E 's/([^ ]+) ([^ ]+) ([^ ]+) ([^ ]+) ([^ ]+)/\1_\2\3\4[\5]/' \
-      | sed -E 's/,/;/g' \
-      | tr '\n' ',' \
-      | sed -E 's/,$//' \
-      | tee ANNOTATED_AMR_SUMMARY_STRING
-
-    echo "PASS" | tee CHROQUETAS_STATUS
+      # extract AMR summary annotated w/fungicide resistance
+      # e.g. <GENE>_<REF_POSITION><AA_CHANGE><QUERY_POSITION>[<FUNGICIDE_RESISTANCE1>;<FUNGICIDE_RESISTANCEn>]
+      tail -n+2 chroquetas_out/~{samplename}.ChroQueTaS.AMR_summary.txt \
+        | awk '{ print $1, $4, $3, $5, $6 }' \
+        | sed -E 's/([^ ]+) ([^ ]+) ([^ ]+) ([^ ]+) ([^ ]+)/\1_\2\3\4[\5]/' \
+        | sed -E 's/,/;/g' \
+        | tr '\n' ',' \
+        | sed -E 's/,$//' \
+        | tee ANNOTATED_AMR_SUMMARY_STRING
+  
+      echo "PASS" | tee CHROQUETAS_STATUS
+    # check if no AMR genes found
+    elif [ -f chroquetas_out/${based_name}.ChroQueTaS.AMR_summary.txt ]; then
+      # if no AMR genes found, create empty output files
+      mv chroquetas_out/${based_name}.ChroQueTaS.AMR_summary.txt chroquetas_out/~{samplename}.ChroQueTaS.AMR_summary.txt
+      echo -e "No AMR genes found" > chroquetas_out/~{samplename}.ChroQueTaS.AMR_stats.txt
+      echo "" | tee AMR_SUMMARY_STRING
+      echo "" | tee ANNOTATED_AMR_SUMMARY_STRING
+      echo "PASS" | tee CHROQUETAS_STATUS
+    else
+      echo "ERROR: Missing output files" | tee CHROQUETAS_STATUS
+    fi
   fi
   >>>
   output {
